requirements.txt
discord.py>=2.3.0
requests>=2.31.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
